# feat(slices): add in-memory caching for slice content and stitched output