This PR adds two simple in-memory caches to speed up the HTML stitching process:

sliceCache - stores the raw HTML of each slice after it's read from disk

stitchedSliceCache - stores the fully-stitched version of each slice to avoid re-processing

With 5,000+ pages being stitched, we were re-reading and re-stitching the same slices many times. Most of our slices are reused heavily, so caching makes a big difference.

On local benchmarks, this reduced stitching time for 5000 pages, 8 slices into every page from ~300s to ~80s.

next steps:
by introducing workers, cold builds speeds improve from ~300s to ~35s