@kdichev 2025-07-27
From my perspective, I’m working with a fairly large Gatsby site, around 5,000 active pages (originally 25,000, but I’ve removed outdated content). Each page includes 8 unique slices.

Before making changes, the slice processing step alone took about 300 seconds, which was surprisingly close to the time required for optimizing 6,000 images. I reviewed the code and introduced some performance improvements that brought the slice processing time down from ~300s to ~80s.

I also tested whether the regex logic was a bottleneck, but replacing it with a custom parser didn’t yield any speed gains. However, moving the slice queue from fastq to worker threads brought cold build time down further to 35s, and hot builds (with slice changes) to around 50s.

I noticed that my machine’s build resources weren’t being fully utilized before, but after introducing workers, usage hit full capacity.

This could be a solid further general improvement. Let me know if this is something you'd consider viable, and I’ll be happy to create add it.