@kdichev 2025-08-13
@pieh This is reasonable feedback, thanks for noting the memory issue! I honestly didn’t think about that at all. I’ve got a fairly powerful machine, so I guess it hasn’t been high on my list of concerns, heh. I’ll work on the suggestions and see how it goes. If there’s already something similar in the repo, and it’s not a burden, I’d appreciate a link so I can take inspiration and stay aligned with accepted practices here.